更新版本和更新代码

+++ Managed/main/Pickupable.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using FMOD.Studio;
+using KSerialization;
 using STRINGS;
 using UnityEngine;
 
@@ -63,6 +64,9 @@ public class Pickupable : Workable, IHasSortOrder
 
 	public const float WorkTime = 1.5f;
 
+	[SerializeField]
+	private int _sortOrder;
+
 	[NonSerialized]
 	[MyCmpReq]
 	public KPrefabID KPrefabID;
@@ -115,6 +119,9 @@ public class Pickupable : Workable, IHasSortOrder
 
 	private int nextTicketNumber;
 
+	[Serialize]
+	public bool deleteOffGrid = true;
+
 	private List<Reservation> reservations = new List<Reservation>();
 
 	private HandleVector<int>.Handle solidPartitionerEntry;
@@ -159,8 +166,14 @@ public class Pickupable : Workable, IHasSortOrder
 
 	public int sortOrder
 	{
-		get;
-		set;
+		get
+		{
+			return _sortOrder;
+		}
+		set
+		{
+			_sortOrder = value;
+		}
 	}
 
 	public Storage storage
@@ -386,7 +399,7 @@ public class Pickupable : Workable, IHasSortOrder
 	{
 		base.OnSpawn();
 		int num = Grid.PosToCell(this);
-		if (!Grid.IsValidCell(num))
+		if (!Grid.IsValidCell(num) && deleteOffGrid)
 		{
 			base.gameObject.DeleteObject();
 			return;
@@ -504,7 +517,7 @@ public class Pickupable : Workable, IHasSortOrder
 		{
 			Vector2 vector = new Vector2(-0.1f * (float)Grid.WidthInCells, 1.1f * (float)Grid.WidthInCells);
 			Vector2 vector2 = new Vector2(-0.1f * (float)Grid.HeightInCells, 1.1f * (float)Grid.HeightInCells);
-			if (position.x < vector.x || vector.y < position.x || position.y < vector2.x || vector2.y < position.y)
+			if (deleteOffGrid && (position.x < vector.x || vector.y < position.x || position.y < vector2.x || vector2.y < position.y))
 			{
 				this.DeleteObject();
 			}
@@ -599,7 +612,7 @@ public class Pickupable : Workable, IHasSortOrder
 			return false;
 		}
 		Absorb(other);
-		if (!hide_effects && EffectPrefabs.Instance != null)
+		if (!hide_effects && EffectPrefabs.Instance != null && !storage)
 		{
 			Vector3 position = base.transform.GetPosition();
 			position.z = Grid.GetLayerZ(Grid.SceneLayer.Front);
@@ -678,7 +691,12 @@ public class Pickupable : Workable, IHasSortOrder
 
 	private void RefreshStorageTags(object data = null)
 	{
-		if (data is Storage || (data != null && (bool)data))
+		bool flag = data is Storage || (data != null && (bool)data);
+		if (flag && data is Storage && ((Storage)data).gameObject == base.gameObject)
+		{
+			return;
+		}
+		if (flag)
 		{
 			KPrefabID.AddTag(GameTags.Stored);
 			if ((object)storage == null || !storage.allowItemRemoval)
@@ -755,6 +773,12 @@ public class Pickupable : Workable, IHasSortOrder
 		base.gameObject.transform.rotation = Quaternion.identity;
 		RegisterListeners();
 		component.GetBatchInstanceData().ClearOverrideTransformMatrix();
+	}
+
+	public void UpdateCachedCellFromStoragePosition()
+	{
+		Debug.Assert(storage != null, "Only call UpdateCachedCellFromStoragePosition on pickupables in storage!");
+		UpdateCachedCell(Grid.PosToCell(storage));
 	}
 
 	private void UpdateCachedCell(int cell)
